K8S shell部署调整

+++ _posts/Work_Notes/部署shell/set_Yum.sh
@@ -5,16 +5,24 @@
 # --------LC---------
 if sudo rpm -qa | grep wget;
 then
-    sudo cp /etc/yum.repos.d/CentOS-Base.repo /etc/yum.repos.d/CentOS-Base.repo.back
-    sudo wget -O /etc/yum.repos.d/CentOS-Base.repo https://mirrors.aliyun.com/repo/Centos-7.repo
-    sudo yum clean all && sudo yum makecache
+    if [ -e '/etc/yum.repos.d/CentOS-Base.repo.back' ]
+    then
+      echo '已更新，无需重复操作'
+      exit 1
+    else
+      sudo cp /etc/yum.repos.d/CentOS-Base.repo /etc/yum.repos.d/CentOS-Base.repo.back
+      sudo wget -O /etc/yum.repos.d/CentOS-Base.repo https://mirrors.aliyun.com/repo/Centos-7.repo
+      sudo yum clean all && sudo yum makecache
+    fi
 else
-    if [ sudo yum install -y wget = 1 ]
-	then
-        sudo cp /etc/yum.repos.d/CentOS-Base.repo /etc/yum.repos.d/CentOS-Base.repo.back
-        sudo wget -O /etc/yum.repos.d/CentOS-Base.repo https://mirrors.aliyun.com/repo/Centos-7.repo
-        sudo yum clean all && sudo yum makecache
-	else
-		echo '错误！请检查网络'
-	fi
+    sudo yum install -y wget
+    if sudo rpm -qa | grep wget
+	  then
+      sudo cp /etc/yum.repos.d/CentOS-Base.repo /etc/yum.repos.d/CentOS-Base.repo.back
+      sudo wget -O /etc/yum.repos.d/CentOS-Base.repo https://mirrors.aliyun.com/repo/Centos-7.repo
+      sudo yum clean all && sudo yum makecache
+	  else
+		  echo '错误！请检查网络'
+		  exit 1
+	  fi
 fi
